# Optimize AsyncJsonReferenceVisitorBase.Visit